files/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
breadability==0.1.20
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
lxml==4.9.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.26.5
pycountry==22.3.5
python-dotenv==0.21.1
regex==2022.10.31
requests==2.28.2
sumy==0.11.0
tinysegmenter==0.4
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
